+++ Week2/src/context/ToDo/ToDoContext.tsx
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+import { IToDoContext } from "../../types/todo";
+
+export const ToDoContext = createContext<IToDoContext | undefined>(undefined);
