+++ Week2/src/context/ToDo/ToDoProvider.tsx
@@ -1,16 +1,6 @@
-/* eslint-disable react-refresh/only-export-components */
-import { createContext, PropsWithChildren, useContext, useState } from "react";
-import { ITodo } from "../types/todo";
-
-interface IToDoContext {
-  toDos: ITodo[];
-  doneToDos: ITodo[];
-  addToDo: (text: string) => void;
-  handleCompleteBtn: (todo: ITodo) => void;
-  handleDeleteBtn: (todo: ITodo) => void;
-}
-
-export const ToDoContext = createContext<IToDoContext | undefined>(undefined);
+import { PropsWithChildren, useState } from "react";
+import { ITodo } from "../../types/todo";
+import { ToDoContext } from "./todoContext";
 
 export const ToDoProvider = ({ children }: PropsWithChildren) => {
   const [toDos, setToDos] = useState<ITodo[]>([]);
@@ -43,16 +33,4 @@ export const ToDoProvider = ({ children }: PropsWithChildren) => {
       {children}
     </ToDoContext.Provider>
   );
-};
-
-export const useTodo = () => {
-  const context = useContext(ToDoContext);
-  // 컨텍스트가 없는 경우 처리
-  if (!context) {
-    throw new Error(
-      "useToDo는 반드시 ToDoProvider 내부에서 사용되어야 합니다."
-    );
-  }
-  // 컨텍스트가 있는 경우
-  return context;
 };
